move Attribute to datastructures folder
