+++ SOMcreator/__init__.py
@@ -1,8 +1,8 @@
-from .datastructure import Project, Object, PropertySet
-
-from .classes import Attribute, Aggregation, UseCase, Phase
-from .external_software import desite, allplan, revit, vestra, card1
 from .constants import json_constants, value_constants
+from .datastructure import Project, Object, PropertySet, Attribute
+
+from .classes import Aggregation, UseCase, Phase
+from .external_software import desite, allplan, revit, vestra, card1
 from .tools import merge_projects
 from .core import export_excel as excel
