+++ SOMcreator/datastructure/__init__.py
@@ -1,3 +1,4 @@
 from .project import Project
 from .object_ import Object
 from .propertyset import PropertySet
+from .attribute import Attribute
